rename nrf_cloud_provision.py to nrf_cloud_onboard.py
IRIS-7817
